# contracts/openapi.json
{
  "components": {
    "schemas": {
      "ActiveFamily": {
        "properties": {
          "family_id": {
            "format": "uuid",
            "title": "Family Id",
            "type": "string"
          }
        },
        "required": [
          "family_id"
        ],
        "title": "ActiveFamily",
        "type": "object"
      },
      "AuthFailure": {
        "properties": {
          "reason": {
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "AuthFailure",
        "type": "object"
      },
...
      "FamilyOut": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        },
        "required": [
          "id",
          "name",
          "role"
        ],
        "title": "FamilyOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HandshakeExchange": {
        "properties": {
          "verifier": {
            "maxLength": 128,
            "minLength": 43,
            "title": "Verifier",
            "type": "string"
          }
        },
        "required": [
          "verifier"
        ],
        "title": "HandshakeExchange",
        "type": "object"
      },
      "HandshakeStart": {
        "properties": {
          "anonymous_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Anonymous Id"
          },
          "challenge": {
            "maxLength": 43,
            "minLength": 43,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Challenge",
            "type": "string"
          },
          "utm": {
            "additionalProperties": {
              "type": "string"
            },
            "maxProperties": 10,
            "title": "Utm",
            "type": "object"
          }
        },
        "required": [
          "challenge"
        ],
        "title": "HandshakeStart",
        "type": "object"
      },
      "HandshakeStarted": {
        "properties": {
          "deep_link": {
            "title": "Deep Link",
            "type": "string"
          },
          "expires_at": {
            "title": "Expires At",
            "type": "string"
          },
          "nonce": {
            "title": "Nonce",
            "type": "string"
          },
          "poll_interval_ms": {
            "title": "Poll Interval Ms",
            "type": "integer"
          }
        },
        "required": [
          "nonce",
          "deep_link",
          "expires_at",
          "poll_interval_ms"
        ],
        "title": "HandshakeStarted",
        "type": "object"
      },
...
      "MagicRedeem": {
        "properties": {
          "anonymous_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Anonymous Id"
          },
          "token": {
            "maxLength": 128,
            "minLength": 16,
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token"
        ],
        "title": "MagicRedeem",
        "type": "object"
      },
      "Me": {
        "properties": {
          "active_family_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Family Id"
          },
          "bot_link": {
            "title": "Bot Link",
            "type": "string"
          },
          "families": {
            "items": {
              "$ref": "#/components/schemas/FamilyOut"
            },
            "title": "Families",
            "type": "array"
          },
          "onboarding_completed": {
            "title": "Onboarding Completed",
            "type": "boolean"
          },
          "user": {
            "$ref": "#/components/schemas/UserOut"
          }
        },
        "required": [
          "user",
          "onboarding_completed",
          "active_family_id",
          "families",
          "bot_link"
        ],
        "title": "Me",
        "type": "object"
      },
      "MeUpdate": {
        "properties": {
          "onboarding_completed": {
            "const": true,
            "title": "Onboarding Completed",
            "type": "boolean"
          }
        },
        "required": [
          "onboarding_completed"
        ],
        "title": "MeUpdate",
        "type": "object"
      },
      "Pending": {
        "properties": {
          "status": {
            "const": "pending",
            "default": "pending",
            "title": "Status",
            "type": "string"
          }
        },
        "title": "Pending",
        "type": "object"
      },
...
      "Role": {
        "enum": [
          "owner",
          "adult",
          "teen",
          "child",
          "caregiver"
        ],
        "title": "Role",
        "type": "string"
      },
      "Session": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "token_type": {
            "const": "bearer",
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "expires_in",
          "user_id"
        ],
        "title": "Session",
        "type": "object"
      },
...
      "UserOut": {
        "properties": {
          "first_name": {
            "title": "First Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "first_name"
        ],
        "title": "UserOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "kainem API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/auth/logout": {
      "post": {
        "operationId": "logout_v1_auth_logout_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "kn_refresh",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kn Refresh"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Logout",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/magic": {
      "post": {
        "description": "The single-use link from the bot message (\"Вот пространство семьи — <link>\").",
        "operationId": "redeem_magic_link_v1_auth_magic_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-client-platform",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Client-Platform"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MagicRedeem"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            },
            "description": "Successful Response"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthFailure"
                }
              }
            },
            "description": "Gone"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Redeem Magic Link",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "description": "Rotates the refresh cookie. Reusing an old one revokes the whole login (theft detection).",
        "operationId": "refresh_v1_auth_refresh_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "kn_refresh",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kn Refresh"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthFailure"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/tg-handshake": {
      "post": {
        "description": "Step 1: the PWA gets a nonce and opens `t.me/<bot>?start=login_<nonce>`.",
        "operationId": "start_handshake_v1_auth_tg_handshake_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-client-platform",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Client-Platform"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HandshakeStart"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HandshakeStarted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Handshake",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/tg-handshake/{nonce}/exchange": {
      "post": {
        "description": "Step 2 (polled): 202 until the user presses Start in the bot, then a session, exactly once.",
        "operationId": "exchange_handshake_v1_auth_tg_handshake__nonce__exchange_post",
        "parameters": [
          {
            "in": "path",
            "name": "nonce",
            "required": true,
            "schema": {
              "title": "Nonce",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HandshakeExchange"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            },
            "description": "Successful Response"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pending"
                }
              }
            },
            "description": "Accepted"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthFailure"
                }
              }
            },
            "description": "Forbidden"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthFailure"
                }
              }
            },
            "description": "Gone"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Exchange Handshake",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/v1/me": {
      "get": {
        "operationId": "get_me_v1_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Me"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Me",
        "tags": [
          "me"
        ]
      },
      "patch": {
        "operationId": "update_me_v1_me_patch",
        "parameters": [
          {
            "in": "header",
            "name": "x-client-platform",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Client-Platform"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Me"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Me",
        "tags": [
          "me"
        ]
      }
    },
    "/v1/me/active-family": {
      "put": {
        "description": "Switch the family shown in the app (a user can belong to several families).",
        "operationId": "set_active_family_v1_me_active_family_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActiveFamily"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Me"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "description": "not a member"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Active Family",
        "tags": [
          "me"
        ]
      }
    }
  }
}